Revert "Union de master a cambios de marzo"
