+++ src/icaro/aplicaciones/agentes/componentesInternos/movimientoCtrl/imp/MaquinaEstadoMovimientoCtrl.java
@@ -54,11 +54,11 @@ public class MaquinaEstadoMovimientoCtrl {
 	public ItfProcesadorObjetivos itfProcObjetivos;
 	protected HebraMonitorizacionLlegada monitorizacionLlegadaDestino;
 	ItfUsoRecursoVisualizadorEntornosSimulacion itfUsoRecVisEntornosSimul;
-	
+	private static ArrayList<LineaObstaculo> obstaculos;
 	private ArrayList<LineaObstaculo> obstaculosDescubiertos;
 
 	public  MaquinaEstadoMovimientoCtrl (){
-		
+		this.obstaculos=VisorEscenariosRosace.getObstaculos();
 		this.obstaculosDescubiertos = new ArrayList<LineaObstaculo>();
 		estadosCreados = new EnumMap<EstadoMovimientoRobot, EstadoAbstractoMovRobot>(EstadoMovimientoRobot.class) ;
 	}
@@ -185,6 +185,15 @@ public class MaquinaEstadoMovimientoCtrl {
 	public synchronized void imposibleAvanzarADestino(){
 		estadoActual = this.cambiarEstado(EstadoMovimientoRobot.RobotBloqueado);
 	}
+
+	public static synchronized boolean checkObstaculo(Coordinate coor){
+		for(LineaObstaculo obs:obstaculos){
+			if(obs.compruebaCoordenada(coor))return true;
+		}
+		return false;
+	}
+
+
 	public synchronized Coordinate getCoordenadasActuales() {
 		return this.robotposicionActual;
 	}
@@ -206,5 +215,15 @@ public class MaquinaEstadoMovimientoCtrl {
 
 
 	
-	
+	public LineaObstaculo getObstaculo(Coordinate coordinate) {
+		for(LineaObstaculo obs:obstaculos){
+			if(obs.compruebaCoordenada(coordinate))return obs;
+		}
+		return null;
+	}
+
+	public void actualizarObstaculos() {
+		obstaculos = VisorEscenariosRosace.getObstaculos();
+		
+	}
 }
